7 funkcji na 10 gotowych
