+++ lista3/src/funkcje/get_entries_by_addr.py
@@ -1,15 +1,4 @@
-def is_dot_decimal(num):
-    parts = str(num).split('.')
-
-    if len(parts) != 4:
-        return False
-
-    for part in parts:
-        if not part.isdigit():
-            if not int(part) >= 0 and int(part) <= 255:
-                return False
-
-    return True
+from lista3.src.funkcje.common import is_dot_decimal
 
 def get_entries_by_addr(log, addr):
     if not isinstance(log, list):
