+++ lista3/src/funkcje/get_top_ips.py
@@ -1,0 +1,9 @@
+from lista3.src.funkcje.common import is_dot_decimal
+
+
+def get_top_ips(log, n=10):
+    ips = []
+    counters = []
+    for tup in log:
+        ips.append(tup[2])
+        is_dot_decimal(tup[4])
